+++ logistic-service/src/main/java/com/sparta/blackyolk/logistic_service/hubroute/framework/web/dto/HubRouteCreateResponse.java
@@ -2,15 +2,14 @@ package com.sparta.blackyolk.logistic_service.hubroute.framework.web.dto;
 
 import com.sparta.blackyolk.logistic_service.hubroute.application.domain.HubRoute;
 import com.sparta.blackyolk.logistic_service.hubroute.data.vo.HubRouteStatus;
-import java.math.BigDecimal;
 
 public record HubRouteCreateResponse(
     String hubRouteId,
     String departureHubName,
     String arrivalHubName,
     HubRouteStatus hubRouteStatus,
-    BigDecimal distance,
-    Integer duration
+    Long distance,
+    Long duration
 ) {
 
     public static HubRouteCreateResponse toDTO(
